backend/requirements.txt
# Core FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI / Computer Vision
insightface==0.7.3
onnxruntime==1.17.0
opencv-python==4.9.0.80
scipy==1.12.0
numpy==1.26.3

# Utilities
pydantic==2.5.3
python-dotenv==1.0.1

# For GPU support (optional - install separately)
# onnxruntime-gpu==1.17.0
